+++ node/resolvers/directives.ts
@@ -1,6 +1,6 @@
 import { AuditAccess } from './directives/auditAccess'
-import { CheckAdminAccess } from './directives/checkAdminAccess'
-import { CheckUserAccess } from './directives/checkUserAccess'
+import { ValidateAdminUserAccess } from './directives/validateAdminUserAccess'
+import { ValidateStoreUserAccess } from './directives/validateStoreUserAccess'
 import { WithPermissions } from './directives/withPermissions'
 import { WithSegment } from './directives/withSegment'
 import { WithSession } from './directives/withSession'
@@ -9,7 +9,7 @@ export const schemaDirectives = {
   withPermissions: WithPermissions as any,
   withSession: WithSession as any,
   withSegment: WithSegment as any,
-  checkAdminAccess: CheckAdminAccess as any,
+  validateAdminUserAccess: ValidateAdminUserAccess as any,
+  validateStoreUserAccess: ValidateStoreUserAccess as any,
   auditAccess: AuditAccess as any,
-  checkUserAccess: CheckUserAccess as any,
 }