+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [B2BTEAM-3133] Add `view-seller` permission to access Seller Register API endpoints. This is required to comply with VTEX's new granular permissions security policy.
+
 ## [4.0.3] - 2026-01-29
 
 ### Changed