+++ CHANGELOG.md
@@ -6,12 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [4.0.5] - 2026-03-04
+## [Unreleased]
 
 ### Changed
 - Atualização de versão.
-
-## [Unreleased]
 
 ## [4.0.4] - 2026-02-04