+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
-- Atualização de versão.
+- Update documentation version.
 
 ## [4.0.4] - 2026-02-04