+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
 
 ### Fixed
@@ -16,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added unitMultiplier to QuoteItem schema and typings
 - Fixed subtotal calculations to include unitMultiplier
 - Updated quote application logic to support unitMultiplier
+
+### Changed
+- Update documentation version.
 
 ## [4.0.4] - 2026-02-04